% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_mob_tps.R
\name{ind.mob.tps}
\alias{ind.mob.tps}
\title{Calcul des indices de mobilite dans le temps}
\usage{
ind.mob.tps(debut, fin, intervalle, var.t, var.tpk, data, periode = "annee",
  poids = NULL, mobilite = TRUE)
}
\arguments{
\item{debut}{indique la premiere periode a laquelle les indices doivent etre calcules}

\item{fin}{indique la periode a laquelle les calcul doivent s'arreter}

\item{intervalle}{indique a des intervalles de combien de periode la fonction doit calculer 
les indices.}

\item{var.t}{une variable discrete correspondant aux etats possibles a la periode t}

\item{var.tpk}{une variable discrete correspondant aux etats possibles a la periode t 
plus l'intervalle k}

\item{data}{un dataframe dans lequel sont contenus var.t et var.tpk}

\item{periode}{une variable definissant les periodes}

\item{poids}{une variable contenant les poids a considerer}

\item{mobilite}{un parametre logique indiquant si les indices de mobilite ou d'immobilite
doivent etre retournes}
}
\value{
une liste de 6 elements
\item{CALL}{Appel de la fonction}
\item{data}{Les donnees utilisees pour effectuer le calcul des indices}
\item{cat.var.t}{Categories que prend la variable var.t}
\item{cat.var.tpk}{Categories que prend la variable var.tpk}
\item{Poids}{Nom de la variable "poids" utilisee}
\item{indices}{Le nom des indices calcules}
\item{ind}{La valeur des indices calculees pour chaque periode definie}
}
\description{
\code{ind.mob.tps} calcule differents indices de mobilite pour des donnes longitudinales
selon des periodes d'evaluation a definir
}
\details{
La variable de la periode t est toujours compare a celle de la periode t plus
l'intervalle k, l'intervalle comparatif k est donc defini par les donnees. Le parametre
"intervalle" determine a des intervalles de combien de periode la fonction doit calculer 
les indices.
}
\author{
Nicolas Corneau-Tremblay
}
