% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatMobil.R
\docType{package}
\name{MatMobil-package}
\alias{MatMobil-package}
\alias{MatMobil}
\title{Fonctions pour le calcul de matrices de transition et d'indices de mobilité}
\description{
MatMobil fournit des outils permettant le calcul de matrice de transition et
d'indices de mobilité (et d'immobilité).
}
