% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mob_bas.R
\name{mob.bas}
\alias{mob.bas}
\title{Calcul de la mobilité vers le bas}
\usage{
mob.bas(mat)
}
\arguments{
\item{mat}{une matrice de transition ou un objet "mat.tr"}
}
\value{
une liste de 4 éléments
\item{CALL}{Appel de la fonction}
\item{matrice}{Matrice de transition utilisée}
\item{indice}{Nom de l'indice calculé}
\item{ind}{Valeur de l'indice calculé}
}
\description{
\code{mob.bas} calcule la proportion d'individus ayant descendu de quantile
}
\details{
A compléter
}
\author{
Nicolas Corneau-Tremblay
}
