% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_im.R
\name{ratio.im}
\alias{ratio.im}
\title{Calcul du ratio d'immobilité}
\usage{
ratio.im(mat)
}
\arguments{
\item{mat}{une matrice de transition ou un objet "mat.tr"}
}
\value{
une liste de 4 éléments
\item{CALL}{Appel de la fonction}
\item{matrice}{Matrice de transition utilisée}
\item{indice}{Nom de l'indice calculé}
\item{ind}{Valeur de l'indice calculé}
}
\description{
\code{ratio.im} calcule le ratio d'immobilité d'une matrice de transition
}
\details{
A completer
}
\author{
Nicolas Corneau-Tremblay
}
