% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_mob.R
\name{ratio.mob}
\alias{ratio.mob}
\title{Calcul du ratio de mobilité}
\usage{
ratio.mob(mat)
}
\arguments{
\item{mat}{une matrice de transition ou un objet "mat.tr"}
}
\value{
une liste de 4 éléments
\item{CALL}{Appel de la fonction}
\item{matrice}{Matrice de transition utilisée}
\item{indice}{Nom de l'indice calculé}
\item{ind}{Valeur de l'indice calculé}
}
\description{
\code{ratio.mob} calcule le ratio de mobilité d'une matrice de transition
}
\details{
1 - Proportion d'individu sur la diagonale de la matrice de transition
}
\author{
Nicolas Corneau-Tremblay
}
