% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saut_moy.R
\name{saut.moy}
\alias{saut.moy}
\title{Calcul du saut moyen}
\usage{
saut.moy(mat)
}
\arguments{
\item{mat}{une matrice de transition ou un objet "mat.tr"}
}
\value{
une liste de 4 éléments
\item{CALL}{Appel de la fonction}
\item{matrice}{Matrice de transition utilisée}
\item{indice}{Nom de l'indice calculé}
\item{ind}{Valeur de l'indice calculé}
}
\description{
\code{trace.norm} calcule le saut moyen d'une matrice de transition
}
\details{
Aussi appelé indice de Bartholomew. Somme sur i des pi_i * somme sur j
des p_ij * valeur absolue de i moins j
}
\author{
Nicolas Corneau-Tremblay
}
