% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_norm.R
\name{trace.norm}
\alias{trace.norm}
\title{Calcul de la trace normaliseé}
\usage{
trace.norm(mat)
}
\arguments{
\item{mat}{une matrice de transition ou un objet "mat.tr"}
}
\value{
une liste de 4 éléments
\item{CALL}{Appel de la fonction}
\item{matrice}{Matrice de transition utilisée}
\item{indice}{Nom de l'indice calculé}
\item{ind}{Valeur de l'indice calculé}
}
\description{
\code{trace.norm} calcule la trace normalisée d'une matrice de transition
}
\details{
Nombre d'états de transition possibles - la trace de la matrice de transition (en prob) 
divisé par le nombre d'états de transition possibles - 1 : ((n - trace(mat)) / (n - 1 ))
}
\author{
Nicolas Corneau-Tremblay
}
