% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_mob_tps.R
\name{ind.mob.tps}
\alias{ind.mob.tps}
\title{Calcul des indices de mobilité dans le temps}
\usage{
ind.mob.tps(debut, fin, intervalle, var.t, var.tpk, data, periode = "annee",
  poids = NULL, mobilite = TRUE)
}
\arguments{
\item{debut}{indique la première période à laquelle les indices doivent être calculés}

\item{fin}{indique la période à laquelle les calculs doivent s'arrêter}

\item{intervalle}{indique à des intervalles de combien de périodes la fonction doit calculer 
les indices.}

\item{var.t}{une variable discrète correspondant aux états possibles à la période t}

\item{var.tpk}{une variable discrète correspondant aux états possibles à la période t 
plus l'intervalle k}

\item{data}{un dataframe dans lequel sont contenues var.t et var.tpk}

\item{periode}{une variable définissant les périodes}

\item{poids}{une variable contenant les poids à considerer}

\item{mobilite}{un paramètre logique indiquant si les indices de mobilité ou d'immobilité
doivent être retournés}
}
\value{
une liste de 6 éléments
\item{CALL}{Appel de la fonction}
\item{data}{Données utilisées pour effectuer le calcul des indices}
\item{cat.var.t}{Catégories que prend la variable var.t}
\item{cat.var.tpk}{Catégories que prend la variable var.tpk}
\item{Poids}{Nom de la variable "poids" utilisée}
\item{indices}{Noms des indices calculés}
\item{ind}{Valeur des indices calculés pour chaque période définie}
}
\description{
\code{ind.mob.tps} calcule différents indices de mobilité pour des données longitudinales
selon des périodes d'évaluation à définir
}
\details{
La variable de la période t est toujours comparée à celle de la période t plus
l'intervalle k. L'intervalle comparatif k est donc défini par les données. Le paramètre
"intervalle" détermine à des intervalles de combien de périodes la fonction doit calculer 
les indices.
}
\author{
Nicolas Corneau-Tremblay
}
