% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_tr.R
\name{mat.tr}
\alias{mat.tr}
\title{Calcul d'une matrice de transition}
\usage{
mat.tr(var.t, var.tpk, data, poids = NULL, prob = FALSE)
}
\arguments{
\item{var.t}{une variable discrete correspondant aux etats possibles a la periode t}

\item{var.tpk}{une variable discrete correspondant aux etats possibles a la periode t 
plus l'intervalle k}

\item{data}{un dataframe dans lequel sont contenus var.t et var.tpk}

\item{poids}{une variable contenant les poids a considerer}

\item{prob}{un parametre logique indiquant si les cellules de la matrice retournee 
doivent contenir des probabilites ou le nombre d'observations si retrouvant}
}
\value{
une liste de 3 elements
\item{CALL}{Appel de la fonction}
\item{Poids}{Nom de la variable "poids" utilisee}
\item{n.obs}{Nombre d'observations (si "poids" specifie, somme des poids individuels)}
\item{taille.mat}{Taille de la matrice retournee}
\item{matrice}{La matrice de transition calculee}
}
\description{
\code{mat.tr} calcule une matrice de transition
}
\details{
La variable de la periode t est toujours compare a celle de la periode t plus
l'intervalle k, l'intervalle comparatif k est donc defini par les donnees.
}
\author{
Nicolas Corneau-Tremblay
}
