% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_im_aj.R
\name{ratio.im.aj}
\alias{ratio.im.aj}
\title{Calcul du ratio d'immobilite ajuste}
\usage{
ratio.im.aj(mat)
}
\arguments{
\item{mat}{une matrice de transition ou un objet "mat.tr"}
}
\value{
une liste de 4 elements
\item{CALL}{Appel de la fonction}
\item{matrice}{La matrice de transition utilisee}
\item{indice}{Le nom de l'indice calcule}
\item{ind}{La valeur de l'indice calculee}
}
\description{
\code{ratio.im.aj} calcule le ratio d'immobilite ajuste d'une matrice de transition
}
\details{
Proportion d'individu sur la diagonale de la matrice de transition plus ceux dans les 
cellules adjacentes a la diagonale
}
\author{
Nicolas Corneau-Tremblay
}
