% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_im.R
\name{ratio.im}
\alias{ratio.im}
\title{Calcul le ratio d'immobilite}
\usage{
ratio.im(mat)
}
\arguments{
\item{mat}{une matrice de transition ou un objet "mat.tr"}
}
\value{
une liste de 4 elements
\item{CALL}{Appel de la fonction}
\item{matrice}{La matrice de transition utilisee}
\item{indice}{Le nom de l'indice calcule}
\item{ind}{La valeur de l'indice calculee}
}
\description{
\code{ratio.im} calcule le ratio d'immobilite d'une matrice de transition
}
\details{
A completer
}
\author{
Nicolas Corneau-Tremblay
}
