% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_mob.R
\name{ratio.mob}
\alias{ratio.mob}
\title{Calcul du ratio de mobilite}
\usage{
ratio.mob(mat)
}
\arguments{
\item{mat}{une matrice de transition ou un objet "mat.tr"}
}
\value{
une liste de 4 elements
\item{CALL}{Appel de la fonction}
\item{matrice}{La matrice de transition utilisee}
\item{indice}{Le nom de l'indice calcule}
\item{ind}{La valeur de l'indice calculee}
}
\description{
\code{ratio.mob} calcule le ratio de mobilite d'une matrice de transition
}
\details{
1 - Proportion d'individu sur la diagonale de la matrice de transition
}
\author{
Nicolas Corneau-Tremblay
}
