% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saut_moy.R
\name{saut.moy}
\alias{saut.moy}
\title{Calcul du saut moyen}
\usage{
saut.moy(mat)
}
\arguments{
\item{mat}{une matrice de transition ou un objet "mat.tr"}
}
\value{
une liste de 4 elements
\item{CALL}{Appel de la fonction}
\item{matrice}{La matrice de transition utilisee}
\item{indice}{Le nom de l'indice calcule}
\item{ind}{La valeur de l'indice calculee}
}
\description{
\code{trace.norm} calcule le saut moyen d'une matrice de transition
}
\details{
Aussi appele indice de Bartholomew. Somme sur i des pi_i * somme sur j
des p_ij * valeur absolue de i moins j
}
\author{
Nicolas Corneau-Tremblay
}
