% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_norm.R
\name{trace.norm}
\alias{trace.norm}
\title{Calcul de la trace normalisee}
\usage{
trace.norm(mat)
}
\arguments{
\item{mat}{une matrice de transition ou un objet "mat.tr"}
}
\value{
une liste de 4 elements
\item{CALL}{Appel de la fonction}
\item{matrice}{La matrice de transition utilisee}
\item{indice}{Le nom de l'indice calcule}
\item{ind}{La valeur de l'indice calculee}
}
\description{
\code{trace.norm} calcule la trace normalisee d'une matrice de transition
}
\details{
Nombre d'etats de transition possibles - la trace de la matrice de transition (en prob) 
divise par le nombre d'etats de transition possibles - 1 : ((n - trace(mat)) / (n - 1 ))
}
\author{
Nicolas Corneau-Tremblay
}
